GithubService: refactor
